* rocDecode/Bitstream reader: Moved MD5 code into single header file. Added changes based on review comments.
