+++ samples/videoDecodeRGB/videodecrgb.cpp
@@ -42,7 +42,7 @@ THE SOFTWARE.
 #include "video_demuxer.h"
 #include "roc_video_dec.h"
 #include "video_post_process.h"
-#include "roc_md5.h"
+#include "md5.h"
 
 std::vector<std::string> st_output_format_name = {"native", "bgr", "bgr48", "rgb", "rgb48", "bgra", "bgra64", "rgba", "rgba64"};
